move GetBrightmapFor to archivecollection
