+++ Core/Resources/Archives/Collection/ArchiveCollection.cs
@@ -26,6 +26,7 @@ using Helion.Resources.Definitions.Locks;
 using Helion.Resources.Definitions.MapInfo;
 using Helion.Resources.Definitions.SoundInfo;
 using Helion.Resources.Definitions.Texture;
+using Helion.Resources.Definitions.Zdoom;
 using Helion.Resources.Images;
 using Helion.Resources.IWad;
 using Helion.Resources.Textures;
@@ -699,5 +700,37 @@ public class ArchiveCollection : IResources, IPathResolver
         }
 
         return (iwad, pwads);
+    }
+
+    public BrightmapDefinition? GetBrightmapFor(string textureName, ResourceNamespace textureNamespace)
+    {
+        // TODO: cache
+        System.Diagnostics.Debug.WriteLine($"lookup for {textureName}");
+        var bmapsDef = Definitions.GldefsDefinition.BrightMaps;
+        var brightmapsOfType = textureNamespace switch
+        {
+            ResourceNamespace.Flats => bmapsDef.Flats,
+            ResourceNamespace.Sprites => bmapsDef.Sprites,
+            ResourceNamespace.Textures => bmapsDef.Textures,
+            _ => null
+        };
+
+        // brightmaps can optionally apply to only an IWAD or specific WAD
+        var sourceWad = Entries.FindByNamespace(textureName, textureNamespace, noFallback: true)?.Parent;
+        bool sourceIsIwad = sourceWad?.ArchiveType == ArchiveType.IWAD;
+        string? sourceWadHash = sourceWad?.MD5;
+
+        BrightmapDefinition? brightmap = brightmapsOfType?.FirstOrDefault(x => (
+            x.TargetTexture.EqualsIgnoreCase(textureName)
+            && (
+                (!x.IwadOnly && x.SpecificWadMd5 == null)
+                || (x.IwadOnly && sourceIsIwad)
+                || (x.SpecificWadMd5 == sourceWadHash)
+            )
+        ));
+        if (brightmap == null && bmapsDef.Auto.TryGetValue(textureName, out BrightmapDefinition? val))
+            brightmap = val;
+
+        return brightmap;
     }
 }
